refactor: extract shared OpenAPI helpers into engine/api/openapi-utils
- endpointTable, addDiscriminators, injectConnectorSchemas, connectorSchemaName
  moved to apps/engine/src/api/openapi-utils.ts
- Engine app.ts imports from local openapi-utils.js (removes 100 lines)
- Service app.ts imports endpointTable + addDiscriminators from
  @stripe/sync-engine/api/openapi-utils (removes local duplicates)
- New export: @stripe/sync-engine/api/openapi-utils

Committed-By-Agent: claude
